version. test

+++ tz-lego.sh
@@ -55,6 +55,31 @@ function auto_reload() {
     fi
 }
 function upkeep() {
+    local_version="1.1.2"
+    SCRIPT_PATH="$(readlink -f "$BASH_SOURCE")"
+    version_gt() {
+    [ "$1" != "$2" ] && \
+    [ "$(printf "%s\n%s\n" "$2" "$1" | sort -V | head -n1)" = "$2" ]
+}
+    remote_version=$(curl -fsSL "https://raw.githubusercontent.com/janniktaulan/lego-testing/main/version.txt"  | tr -d '\r' | tr -d '\n' | xargs)
+    if [ -z "$remote_version" ]; then
+        echo "Error fetching remote version."
+        exit 1
+    fi
+    if version_gt "$remote_version" "$local_version"; then
+        read -n 1 -p "New version found: $remote_version. Do you want to update? (y/n): " update_choice
+        if [[ "$update_choice" == "y" ]]; then
+            curl -fsSL "https://raw.githubusercontent.com/janniktaulan/lego-testing/main/tz-lego.sh" \
+        -o "$SCRIPT_PATH.tmp" || exit 1
+
+        mv "$SCRIPT_PATH.tmp" "$SCRIPT_PATH"
+        chmod +x "$SCRIPT_PATH"
+            echo ""
+            echo "Update done! Please run tz-bot again."
+            exit 0
+        fi
+    fi
+
     if ! command -v tz-bot >/dev/null 2>&1; then
         sudo mkdir -p /usr/local/bin
         if sudo mv /tmp/tz-bot /usr/local/bin/tz-bot; then
